feat: #7 Coordinator Interface 수정
