+++ WalWal/Coordinators/SampleApp/SampleAppCoordinator/Interface/SampleAppCoordinator.swift
@@ -9,6 +9,15 @@
 import UIKit
 import Utility
 
+public enum SampleAppCoordinatorAction: ParentAction {
+  case never
+}
+
+public enum SampleAppCoordinatorFlow: CoordinatorFlow {
+  case startAuth
+  case startHome
+}
+
 public protocol SampleAppCoordinator: CoordinatorType {
 
 }
